Fix ALU timing
Divide the clock signal into twice slower one, and add another time
signal that rises with half-phase latency (i.e. the raising edge is in
the middle of the high level of the main clock). This phase-shifted
clock drives the ALU, allowing it to read and write the accumulator
register in the middle of the cycle, thus avoiding oscillations.