Store the Fibonacci calculator in ROM
Store the following program in ROM. The program calculates the Fibonacci
sequence, and stores it in RAM, starting with address 0, onward. It
stops calculating when the Nth Fibonacci number overflows my 8-bit
registers.

```
09  li 1
4f  getacc r7
4e  getacc r6

08  li 0
1e  st r6
88  inc
1f  st r7
88  inc
4d  getacc r5

24  getpc r4
0c  li 4
7c  add r4
4c  getacc r4

57  setacc r7
4a  getacc r2
56  setacc r6
7f  add r7
2b  getst r3

4f  getacc r7
52  setacc r2
4e  getacc r6

55  setacc r5
1f  st r7
88  inc
4d  getacc r5

54  setacc r4
33  setst r3
b8  jno
```

The program was handcrafted, then cross-checked with the disassembler
(as I don't have the assembler yet).