Add docker image to run Clojure REPL
This is meant to be copied to a project dir and executed from there.
Then cider-nrepl is able to see the project files and is generally not
confused

dotlein-profiles.clj is meant to be copied to .lein/profiles.clj under
the project. Should probably bake it into the Dockerfile some day.